Now the classical_damage calculator works with more than one realization
