+++ openquake/commonlib/sourceconverter.py
@@ -1,0 +1,636 @@
+# Copyright (c) 2015, GEM Foundation.
+#
+# OpenQuake is free software: you can redistribute it and/or modify it
+# under the terms of the GNU Affero General Public License as published
+# by the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# OpenQuake is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU Affero General Public License
+# along with OpenQuake.  If not, see <http://www.gnu.org/licenses/>.
+
+import math
+import copy
+import operator
+from itertools import izip
+
+from openquake.hazardlib import geo, mfd, pmf, source
+from openquake.hazardlib.tom import PoissonTOM
+from openquake.commonlib.node import context, striptag
+from openquake.commonlib import valid
+from openquake.commonlib import parallel
+
+# this must stay here for the nrml_converters: don't remove it!
+from openquake.commonlib.obsolete import NrmlHazardlibConverter
+
+# the following is arbitrary, it is used to decide when to parallelize
+# the filtering (MS)
+LOTS_OF_SOURCES_SITES = 1E5  # arbitrary, set by Michele Simionato
+MAGNITUDE_FOR_RUPTURE_SPLITTING = 6.5  # given by Marco Pagani
+# NB: the parameter MAGNITUDE_FOR_RUPTURE_SPLITTING cannot go in a
+# configuration file, otherwise the tests will break by changing it;
+# reason: the numbers in the event based calculators depend on the
+# splitting: different sources => different seeds => different numbers
+
+
+def area_to_point_sources(area_src, area_src_disc):
+    """
+    Split an area source into a generator of point sources.
+
+    MFDs will be rescaled appropriately for the number of points in the area
+    mesh.
+
+    :param area_src:
+        :class:`openquake.hazardlib.source.AreaSource`
+    :param float area_src_disc:
+        Area source discretization step, in kilometers.
+    """
+    mesh = area_src.polygon.discretize(area_src_disc)
+    num_points = len(mesh)
+    area_mfd = area_src.mfd
+
+    if isinstance(area_mfd, mfd.TruncatedGRMFD):
+        new_a_val = math.log10(10 ** area_mfd.a_val / float(num_points))
+        new_mfd = mfd.TruncatedGRMFD(
+            a_val=new_a_val,
+            b_val=area_mfd.b_val,
+            bin_width=area_mfd.bin_width,
+            min_mag=area_mfd.min_mag,
+            max_mag=area_mfd.max_mag)
+    elif isinstance(area_mfd, mfd.EvenlyDiscretizedMFD):
+        new_occur_rates = [float(x) / num_points
+                           for x in area_mfd.occurrence_rates]
+        new_mfd = mfd.EvenlyDiscretizedMFD(
+            min_mag=area_mfd.min_mag,
+            bin_width=area_mfd.bin_width,
+            occurrence_rates=new_occur_rates)
+
+    for i, (lon, lat) in enumerate(izip(mesh.lons, mesh.lats)):
+        pt = source.PointSource(
+            # Generate a new ID and name
+            source_id='%s-%s' % (area_src.source_id, i),
+            name='%s-%s' % (area_src.name, i),
+            tectonic_region_type=area_src.tectonic_region_type,
+            mfd=new_mfd,
+            rupture_mesh_spacing=area_src.rupture_mesh_spacing,
+            magnitude_scaling_relationship=
+            area_src.magnitude_scaling_relationship,
+            rupture_aspect_ratio=area_src.rupture_aspect_ratio,
+            upper_seismogenic_depth=area_src.upper_seismogenic_depth,
+            lower_seismogenic_depth=area_src.lower_seismogenic_depth,
+            location=geo.Point(lon, lat),
+            nodal_plane_distribution=area_src.nodal_plane_distribution,
+            hypocenter_distribution=area_src.hypocenter_distribution,
+            temporal_occurrence_model=area_src.temporal_occurrence_model)
+        pt.trt_model_id = area_src.trt_model_id
+        yield pt
+
+
+def split_fault_source(src):
+    """
+    Generator splitting a fault source into several fault sources,
+    one for each magnitude.
+
+    :param src:
+        an instance of :class:`openquake.hazardlib.source.base.SeismicSource`
+    """
+    i = 0  # split source index
+    max_mag = src.get_min_max_mag()[1]
+    if max_mag > MAGNITUDE_FOR_RUPTURE_SPLITTING:
+        for rupture in src.iter_ruptures():
+            i += 1
+            yield SingleRuptureSource(
+                rupture, '%s-%s' % (src.source_id, i),
+                src.tectonic_region_type, src.trt_model_id)
+    else:  # split on the annual occurrence rates
+        for mag, rate in src.mfd.get_annual_occurrence_rates():
+            if not rate:  # ignore zero occurency rate
+                continue
+            new_src = copy.copy(src)
+            new_src.source_id = '%s-%s' % (src.source_id, i)
+            new_src.mfd = mfd.EvenlyDiscretizedMFD(
+                min_mag=mag, bin_width=src.mfd.bin_width,
+                occurrence_rates=[rate])
+            i += 1
+            yield new_src
+
+
+class SingleRuptureSource(object):
+    """
+    Fake source class used to encapsule the ruptures generated by a
+    ComplexFaultSource.
+
+    :param rupture:
+        an instance of :class:`openquake.hazardlib.source.rupture.
+        ParametricProbabilisticRupture`
+    :param source_id:
+        an ID for the SingleRuptureSource
+    :param tectonic_region_type:
+        the tectonic region type
+    :param trt_model_id:
+        ID of the tectonic region model the source belongs to
+    """
+    def __init__(self, rupture, source_id, tectonic_region_type, trt_model_id):
+        self.rupture = rupture
+        self.source_id = source_id
+        self.tectonic_region_type = tectonic_region_type
+        self.trt_model_id = trt_model_id
+
+    def iter_ruptures(self):
+        """Yield the rupture"""
+        yield self.rupture
+
+    def count_ruptures(self):
+        """Return 1"""
+        return 1
+
+    def filter_sites_by_distance_to_source(self, maxdist, sitecol):
+        """The source has been already filtered, return the sitecol"""
+        return sitecol
+
+
+def split_source(src, area_source_discretization):
+    """
+    Split an area source into point sources and a fault sources into
+    smaller fault sources.
+
+    :param src:
+        an instance of :class:`openquake.hazardlib.source.base.SeismicSource`
+    :param float area_source_discretization:
+        area source discretization
+    """
+    if isinstance(src, source.AreaSource):
+        for s in area_to_point_sources(src, area_source_discretization):
+            yield s
+    elif isinstance(
+            src, (source.SimpleFaultSource, source.ComplexFaultSource)):
+        for s in split_fault_source(src):
+            yield s
+    else:
+        # characteristic and nonparametric sources are not split
+        # since they are small anyway
+        yield src
+
+
+def split_coords_2d(seq):
+    """
+    :param seq: a flat list with lons and lats
+    :returns: a validated list of pairs (lon, lat)
+
+    >>> split_coords_2d([1.1, 2.1, 2.2, 2.3])
+    [(1.1, 2.1), (2.2, 2.3)]
+    """
+    lons, lats = [], []
+    for i, el in enumerate(seq):
+        if i % 2 == 0:
+            lons.append(valid.longitude(el))
+        elif i % 2 == 1:
+            lats.append(valid.latitude(el))
+    return zip(lons, lats)
+
+
+def split_coords_3d(seq):
+    """
+    :param seq: a flat list with lons, lats and depths
+    :returns: a validated list of (lon, lat, depths) triplets
+
+    >>> split_coords_3d([1.1, 2.1, 0.1, 2.3, 2.4, 0.1])
+    [(1.1, 2.1, 0.1), (2.3, 2.4, 0.1)]
+    """
+    lons, lats, depths = [], [], []
+    for i, el in enumerate(seq):
+        if i % 3 == 0:
+            lons.append(valid.longitude(el))
+        elif i % 3 == 1:
+            lats.append(valid.latitude(el))
+        elif i % 3 == 2:
+            depths.append(valid.depth(el))
+    return zip(lons, lats, depths)
+
+
+class RuptureConverter(object):
+    """
+    Convert ruptures from nodes into Hazardlib ruptures.
+    """
+    fname = None  # should be set externally
+
+    def __init__(self, rupture_mesh_spacing, complex_fault_mesh_spacing):
+        self.rupture_mesh_spacing = rupture_mesh_spacing
+        self.complex_fault_mesh_spacing = complex_fault_mesh_spacing
+
+    def convert_node(self, node):
+        """
+        Convert the given rupture node into a hazardlib rupture, depending
+        on the node tag.
+
+        :param node: a node representing a rupture
+        """
+        with context(self.fname, node):
+            convert_rupture = getattr(self, 'convert_' + striptag(node.tag))
+            mag = ~node.magnitude
+            rake = ~node.rake
+            hypocenter = ~node.hypocenter
+        return convert_rupture(node, mag, rake, hypocenter)
+
+    def geo_line(self, edge):
+        """
+        Utility function to convert a node of kind edge
+        into a :class:`openquake.hazardlib.geo.Line` instance.
+
+        :param edge: a node describing an edge
+        """
+        with context(self.fname, edge.LineString.posList) as plist:
+            coords = split_coords_2d(~plist)
+        return geo.Line([geo.Point(*p) for p in coords])
+
+    def geo_lines(self, edges):
+        """
+        Utility function to convert a list of edges into a list of
+        :class:`openquake.hazardlib.geo.Line` instances.
+
+        :param edge: a node describing an edge
+        """
+        lines = []
+        for edge in edges:
+            with context(self.fname, edge):
+                coords = split_coords_3d(~edge.LineString.posList)
+            lines.append(geo.Line([geo.Point(*p) for p in coords]))
+        return lines
+
+    def geo_planar(self, surface):
+        """
+        Utility to convert a PlanarSurface node with subnodes
+        topLeft, topRight, bottomLeft, bottomRight into a
+        :class:`openquake.hazardlib.geo.PlanarSurface` instance.
+
+        :param surface: PlanarSurface node
+        """
+        with context(self.fname, surface):
+            top_left = geo.Point(*~surface.topLeft)
+            top_right = geo.Point(*~surface.topRight)
+            bottom_left = geo.Point(*~surface.bottomLeft)
+            bottom_right = geo.Point(*~surface.bottomRight)
+        return geo.PlanarSurface.from_corner_points(
+            self.rupture_mesh_spacing,
+            top_left, top_right, bottom_right, bottom_left)
+
+    def convert_surfaces(self, surface_nodes):
+        """
+        Utility to convert a list of surface nodes into a single hazardlib
+        surface. There are three possibilities:
+
+        1. there is a single simpleFaultGeometry node; returns a
+           :class:`openquake.hazardlib.geo.simpleFaultSurface` instance
+        2. there is a single complexFaultGeometry node; returns a
+           :class:`openquake.hazardlib.geo.complexFaultSurface` instance
+        3. there is a list of PlanarSurface nodes; returns a
+           :class:`openquake.hazardlib.geo.MultiSurface` instance
+
+        :param surface_nodes: surface nodes as just described
+        """
+        surface_node = surface_nodes[0]
+        if surface_node.tag.endswith('simpleFaultGeometry'):
+            surface = geo.SimpleFaultSurface.from_fault_data(
+                self.geo_line(surface_node),
+                ~surface_node.upperSeismoDepth,
+                ~surface_node.lowerSeismoDepth,
+                ~surface_node.dip,
+                self.rupture_mesh_spacing)
+        elif surface_node.tag.endswith('complexFaultGeometry'):
+            surface = geo.ComplexFaultSurface.from_fault_data(
+                self.geo_lines(surface_node),
+                self.complex_fault_mesh_spacing)
+        else:  # a collection of planar surfaces
+            planar_surfaces = map(self.geo_planar, surface_nodes)
+            surface = geo.MultiSurface(planar_surfaces)
+        return surface
+
+    def convert_simpleFaultRupture(self, node, mag, rake, hypocenter):
+        """
+        Convert a simpleFaultRupture node.
+
+        :param node: the rupture node
+        :param mag: the rupture magnitude
+        :param rake: the rupture rake angle
+        :param hypocenter: the rupture hypocenter
+        """
+        with context(self.fname, node):
+            surfaces = [node.simpleFaultGeometry]
+        rupt = source.rupture.Rupture(
+            mag=mag, rake=rake, tectonic_region_type=None,
+            hypocenter=geo.Point(*hypocenter),
+            surface=self.convert_surfaces(surfaces),
+            source_typology=source.SimpleFaultSource)
+        return rupt
+
+    def convert_complexFaultRupture(self, node, mag, rake, hypocenter):
+        """
+        Convert a complexFaultRupture node.
+
+        :param node: the rupture node
+        :param mag: the rupture magnitude
+        :param rake: the rupture rake angle
+        :param hypocenter: the rupture hypocenter
+        """
+        with context(self.fname, node):
+            surfaces = [node.complexFaultGeometry]
+        rupt = source.rupture.Rupture(
+            mag=mag, rake=rake, tectonic_region_type=None,
+            hypocenter=geo.Point(*hypocenter),
+            surface=self.convert_surfaces(surfaces),
+            source_typology=source.ComplexFaultSource)
+        return rupt
+
+    def convert_singlePlaneRupture(self, node, mag, rake, hypocenter):
+        """
+        Convert a singlePlaneRupture node.
+
+        :param node: the rupture node
+        :param mag: the rupture magnitude
+        :param rake: the rupture rake angle
+        :param hypocenter: the rupture hypocenter
+        """
+        with context(self.fname, node):
+            surfaces = [node.planarSurface]
+        hrupt = source.rupture.Rupture(
+            mag=mag, rake=rake,
+            tectonic_region_type=None,
+            hypocenter=geo.Point(*hypocenter),
+            surface=self.convert_surfaces(surfaces),
+            source_typology=source.NonParametricSeismicSource)
+        return hrupt
+
+    def convert_multiPlanesRupture(self, node, mag, rake, hypocenter):
+        """
+        Convert a multiPlanesRupture node.
+
+        :param node: the rupture node
+        :param mag: the rupture magnitude
+        :param rake: the rupture rake angle
+        :param hypocenter: the rupture hypocenter
+        """
+        with context(self.fname, node):
+            surfaces = list(node.getnodes('planarSurface'))
+        hrupt = source.rupture.Rupture(
+            mag=mag, rake=rake,
+            tectonic_region_type=None,
+            hypocenter=geo.Point(*hypocenter),
+            surface=self.convert_surfaces(surfaces),
+            source_typology=source.NonParametricSeismicSource)
+        return hrupt
+
+
+class SourceConverter(RuptureConverter):
+    """
+    Convert sources from valid nodes into Hazardlib objects.
+    """
+    def __init__(self, investigation_time, rupture_mesh_spacing,
+                 complex_fault_mesh_spacing, width_of_mfd_bin,
+                 area_source_discretization):
+        self.area_source_discretization = area_source_discretization
+        self.rupture_mesh_spacing = rupture_mesh_spacing
+        self.complex_fault_mesh_spacing = complex_fault_mesh_spacing
+        self.width_of_mfd_bin = width_of_mfd_bin
+        self.tom = PoissonTOM(investigation_time)
+
+    def convert_node(self, node):
+        """
+        Convert the given node into a hazardlib source, depending
+        on the node tag.
+
+        :param node: a node representing a source
+        """
+        with context(self.fname, node):
+            convert_source = getattr(self, 'convert_' + striptag(node.tag))
+        return convert_source(node)
+
+    def convert_mfdist(self, node):
+        """
+        Convert the given node into a Magnitude-Frequency Distribution
+        object.
+
+        :param node: a node of kind incrementalMFD or truncGutenbergRichterMFD
+        :returns: a :class:`openquake.hazardlib.mdf.EvenlyDiscretizedMFD.` or
+                  :class:`openquake.hazardlib.mdf.TruncatedGRMFD` instance
+        """
+        with context(self.fname, node):
+            [mfd_node] = [subnode for subnode in node
+                          if subnode.tag.endswith(
+                              ('incrementalMFD', 'truncGutenbergRichterMFD'))]
+            if mfd_node.tag.endswith('incrementalMFD'):
+                return mfd.EvenlyDiscretizedMFD(
+                    min_mag=mfd_node['minMag'], bin_width=mfd_node['binWidth'],
+                    occurrence_rates=~mfd_node.occurRates)
+            elif mfd_node.tag.endswith('truncGutenbergRichterMFD'):
+                return mfd.TruncatedGRMFD(
+                    a_val=mfd_node['aValue'], b_val=mfd_node['bValue'],
+                    min_mag=mfd_node['minMag'], max_mag=mfd_node['maxMag'],
+                    bin_width=self.width_of_mfd_bin)
+
+    def convert_npdist(self, node):
+        """
+        Convert the given node into a Nodal Plane Distribution.
+
+        :param node: a nodalPlaneDist node
+        :returns: a :class:`openquake.hazardlib.geo.NodalPlane` instance
+        """
+        with context(self.fname, node):
+            npdist = []
+            for np in node.nodalPlaneDist:
+                prob, strike, dip, rake = ~np
+                npdist.append((prob, geo.NodalPlane(strike, dip, rake)))
+            return pmf.PMF(npdist)
+
+    def convert_hpdist(self, node):
+        """
+        Convert the given node into a probability mass function for the
+        hypo depth distribution.
+
+        :param node: a hypoDepthDist node
+        :returns: a :class:`openquake.hazardlib.pmf.PMF` instance
+        """
+        with context(self.fname, node):
+            return pmf.PMF([~hd for hd in node.hypoDepthDist])
+
+    def convert_areaSource(self, node):
+        """
+        Convert the given node into an area source object.
+
+        :param node: a node with tag areaGeometry
+        :returns: a :class:`openquake.hazardlib.source.AreaSource` instance
+        """
+        geom = node.areaGeometry
+        coords = split_coords_2d(~geom.Polygon.exterior.LinearRing.posList)
+        polygon = geo.Polygon([geo.Point(*xy) for xy in coords])
+        msr = valid.SCALEREL[~node.magScaleRel]()
+        return source.AreaSource(
+            source_id=node['id'],
+            name=node['name'],
+            tectonic_region_type=node['tectonicRegion'],
+            mfd=self.convert_mfdist(node),
+            rupture_mesh_spacing=self.rupture_mesh_spacing,
+            magnitude_scaling_relationship=msr,
+            rupture_aspect_ratio=~node.ruptAspectRatio,
+            upper_seismogenic_depth=~geom.upperSeismoDepth,
+            lower_seismogenic_depth=~geom.lowerSeismoDepth,
+            nodal_plane_distribution=self.convert_npdist(node),
+            hypocenter_distribution=self.convert_hpdist(node),
+            polygon=polygon,
+            area_discretization=self.area_source_discretization,
+            temporal_occurrence_model=self.tom)
+
+    def convert_pointSource(self, node):
+        """
+        Convert the given node into a point source object.
+
+        :param node: a node with tag pointGeometry
+        :returns: a :class:`openquake.hazardlib.source.PointSource` instance
+        """
+        geom = node.pointGeometry
+        lon_lat = ~geom.Point.pos
+        msr = valid.SCALEREL[~node.magScaleRel]()
+        return source.PointSource(
+            source_id=node['id'],
+            name=node['name'],
+            tectonic_region_type=node['tectonicRegion'],
+            mfd=self.convert_mfdist(node),
+            rupture_mesh_spacing=self.rupture_mesh_spacing,
+            magnitude_scaling_relationship=msr,
+            rupture_aspect_ratio=~node.ruptAspectRatio,
+            upper_seismogenic_depth=~geom.upperSeismoDepth,
+            lower_seismogenic_depth=~geom.lowerSeismoDepth,
+            location=geo.Point(*lon_lat),
+            nodal_plane_distribution=self.convert_npdist(node),
+            hypocenter_distribution=self.convert_hpdist(node),
+            temporal_occurrence_model=self.tom)
+
+    def convert_simpleFaultSource(self, node):
+        """
+        Convert the given node into a simple fault object.
+
+        :param node: a node with tag areaGeometry
+        :returns: a :class:`openquake.hazardlib.source.SimpleFaultSource`
+                  instance
+        """
+        geom = node.simpleFaultGeometry
+        msr = valid.SCALEREL[~node.magScaleRel]()
+        simple = source.SimpleFaultSource(
+            source_id=node['id'],
+            name=node['name'],
+            tectonic_region_type=node['tectonicRegion'],
+            mfd=self.convert_mfdist(node),
+            rupture_mesh_spacing=self.rupture_mesh_spacing,
+            magnitude_scaling_relationship=msr,
+            rupture_aspect_ratio=~node.ruptAspectRatio,
+            upper_seismogenic_depth=~geom.upperSeismoDepth,
+            lower_seismogenic_depth=~geom.lowerSeismoDepth,
+            fault_trace=self.geo_line(geom),
+            dip=~geom.dip,
+            rake=~node.rake,
+            temporal_occurrence_model=self.tom)
+        return simple
+
+    def convert_complexFaultSource(self, node):
+        """
+        Convert the given node into a complex fault object.
+
+        :param node: a node with tag areaGeometry
+        :returns: a :class:`openquake.hazardlib.source.ComplexFaultSource`
+                  instance
+        """
+        geom = node.complexFaultGeometry
+        msr = valid.SCALEREL[~node.magScaleRel]()
+        cmplx = source.ComplexFaultSource(
+            source_id=node['id'],
+            name=node['name'],
+            tectonic_region_type=node['tectonicRegion'],
+            mfd=self.convert_mfdist(node),
+            rupture_mesh_spacing=self.complex_fault_mesh_spacing,
+            magnitude_scaling_relationship=msr,
+            rupture_aspect_ratio=~node.ruptAspectRatio,
+            edges=self.geo_lines(geom),
+            rake=~node.rake,
+            temporal_occurrence_model=self.tom)
+        return cmplx
+
+    def convert_characteristicFaultSource(self, node):
+        """
+        Convert the given node into a characteristic fault object.
+
+        :param node:
+            a node with tag areaGeometry
+        :returns:
+            a :class:`openquake.hazardlib.source.CharacteristicFaultSource`
+            instance
+        """
+        char = source.CharacteristicFaultSource(
+            source_id=node['id'],
+            name=node['name'],
+            tectonic_region_type=node['tectonicRegion'],
+            mfd=self.convert_mfdist(node),
+            surface=self.convert_surfaces(node.surface),
+            rake=~node.rake,
+            temporal_occurrence_model=self.tom)
+        return char
+
+    def convert_nonParametricSeismicSource(self, node):
+        """
+        Convert the given node into a non parametric source object.
+
+        :param node:
+            a node with tag areaGeometry
+        :returns:
+            a :class:`openquake.hazardlib.source.NonParametricSeismicSource`
+            instance
+        """
+        trt = node['tectonicRegion']
+        rup_pmf_data = []
+        for rupnode in node:
+            probs = pmf.PMF(rupnode['probs_occur'])
+            rup = RuptureConverter.convert_node(self, rupnode)
+            rup.tectonic_region_type = trt
+            rup_pmf_data.append((rup, probs))
+        nps = source.NonParametricSeismicSource(
+            node['id'], node['name'], trt, rup_pmf_data)
+        return nps
+
+
+def parse_ses_ruptures(fname):
+    """
+    Convert a stochasticEventSetCollection file into a set of SES,
+    each one containing ruptures with a tag and a seed.
+    """
+    raise NotImplementedError('parse_ses_ruptures')
+
+
+def _filter_sources(sources, sitecol, maxdist):
+    # called by filter_sources
+    srcs = []
+    for src in sources:
+        sites = src.filter_sites_by_distance_to_source(maxdist, sitecol)
+        if sites is not None:
+            srcs.append(src)
+    return srcs
+
+
+def filter_sources(sources, sitecol, maxdist):
+    """
+    Filter a list of hazardlib sources according to the maximum distance.
+
+    :param sources: the original sources
+    :param sitecol: a :class:`openquake.hazardlib.site.SiteCollection` instance
+    :param maxdist: maximum distance
+    :returns: the filtered sources ordered by source_id
+    """
+    if len(sources) * len(sitecol) > LOTS_OF_SOURCES_SITES:
+        # filter in parallel on all available cores
+        sources = parallel.apply_reduce(
+            _filter_sources, (sources, sitecol, maxdist), operator.add, [])
+    else:
+        # few sources and sites, filter sequentially on a single core
+        sources = _filter_sources(sources, sitecol, maxdist)
+    return sorted(sources, key=operator.attrgetter('source_id'))
